requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.23.5
botocore==1.26.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.5
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1
django-appconf==1.0.4
django-cors-headers==3.13.0
django-crispy-forms==1.12.0
django-imagekit==4.0.2
django-on-heroku==1.1.2
django-storages==1.12.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gh==0.0.4
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
oauthlib==3.2.1
pilkit==2.0
Pillow==8.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
smmap==5.0.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
stripe==4.1.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
